Cooking Refactor to CA
